Add files via upload
Just a little command line tool that allows you to generate a .bikey usable on your server from a .bisign file. The bikey is actually stored in bisign files so it can easily be extracted. If you have lost the bikey file to one of your beloved mods, this is for you!

Successfully tested with .bisign files generated from DSSignFile.exe included in the Arma 3 Tools.

Usage: bisign2bikey bisign_path [bikey_destination]

You can also drag-n-drop the bisign onto the EXE, the bikey will be created in the same folder.